@@ -1,138 +1,93 @@
-<?php declare(strict_types=1);
+<?php
 /**
- * Callback Filter Handler for Monolog.
+ * This file is part of the mimmi20/monolog-callbackfilterhandler package.
  *
- * @category Logging
- * @package  monolog-callbackfilterhandler
- * @author   Laurent Laville <pear@laurent-laville.org>
- * @author   Christophe Coevoet
- * @license  http://www.opensource.org/licenses/bsd-license.php  BSD License
+ * Copyright (c) 2022, Thomas Mueller <mimmi20@live.de>
+ * Copyright (c) 2015-2021, Laurent Laville <pear@laurent-laville.org>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
  */
 
-namespace Bartlett\Monolog\Handler;
+declare(strict_types = 1);
 
+/**
+ * Callback Filter Handler for Monolog.
+ */
+
+namespace Mimmi20\Monolog\Handler;
+
+use Closure;
 use Monolog\Handler\AbstractHandler;
 use Monolog\Handler\HandlerInterface;
 use Monolog\Handler\ProcessableHandlerInterface;
-use Monolog\Logger;
-
+use Monolog\Handler\ProcessableHandlerTrait;
+use Monolog\Level;
+use Monolog\LogRecord;
+use Monolog\ResettableInterface;
+use Psr\Log\LogLevel;
 use RuntimeException;
-use function array_shift;
-use function array_unshift;
-use function is_callable;
+
+use function count;
 use function json_encode;
 
 /**
  * Monolog handler wrapper that filters records based on a list of callback functions.
- *
- * @category Logging
- * @package  monolog-callbackfilterhandler
- * @author   Laurent Laville <pear@laurent-laville.org>
- * @author   Christophe Coevoet
- * @license  http://www.opensource.org/licenses/bsd-license.php BSD License
- * @since    Class available since Release 1.0.0
  */
-class CallbackFilterHandler extends AbstractHandler implements ProcessableHandlerInterface
+final class CallbackFilterHandler extends AbstractHandler implements ProcessableHandlerInterface
 {
+    use ProcessableHandlerTrait;
+
     /**
-     * Handler or factory callable($record, $this)
+     * Filters Closure to restrict log records.
      *
-     * @var callable|HandlerInterface
+     * @var Closure[]
+     * @phpstan-var array<int|string, (Closure(LogRecord, int|string|Level): bool)>
      */
-    protected $handler;
+    protected array $filters;
 
     /**
-     * @var int
-     */
-    protected $handlerLevel;
-
-    /**
-     * Filters callable to restrict log records.
+     * @param Closure|HandlerInterface $handler handler or factory Closure($record, $this)
+     * @param Closure[]                $filters A list of filters to apply
+     * @param int|Level|string         $level   The minimum logging level at which this handler will be triggered
+     * @param bool                     $bubble  Whether the messages that are handled can bubble up the stack or not
+     * @phpstan-param (Closure(LogRecord|null, HandlerInterface): HandlerInterface)|HandlerInterface $handler
+     * @phpstan-param value-of<Level::VALUES>|value-of<Level::NAMES>|Level|LogLevel::* $level
      *
-     * @var callable[]
+     * @throws RuntimeException
      */
-    protected $filters;
+    public function __construct(
+        private Closure | HandlerInterface $handler,
+        array $filters,
+        int | string | Level $level = Level::Debug,
+        bool $bubble = true,
+    ) {
+        parent::__construct($level, $bubble);
 
-    /**
-     * Changes to apply on log records, by a stack of callable
-     *
-     * @var callable[]
-     */
-    protected $processors;
-
-    /**
-     * @param callable|HandlerInterface $handler Handler or factory callable($record, $this).
-     * @param callable[]                $filters A list of filters to apply
-     * @param int|string                $level   The minimum logging level at which this handler will be triggered
-     * @param boolean                   $bubble  Whether the messages that are handled can bubble up the stack or not
-     */
-    public function __construct($handler, array $filters, $level = Logger::DEBUG, bool $bubble = true)
-    {
-        if (!$handler instanceof HandlerInterface) {
-            if (!is_callable($handler)) {
-                throw new RuntimeException(
-                    "The given handler (" . json_encode($handler)
-                    . ") is not a callable nor a Monolog\\Handler\\HandlerInterface object"
-                );
-            }
-        }
-
-        $this->handlerLevel = Logger::toMonologLevel($level); // @phpstan-ignore-line
-        parent::__construct($level, $bubble);    // @phpstan-ignore-line
-
-        $this->handler = $handler;
         $this->filters = [];
-        $this->processors = [];
 
         foreach ($filters as $filter) {
-            if (!is_callable($filter)) {
+            if (!$filter instanceof Closure) {
                 throw new RuntimeException(
-                    "The given filter (" . json_encode($filter)
-                    . ") is not a callable object"
+                    'The given filter (' . json_encode($filter) . ') is not a Closure',
                 );
             }
+
             $this->filters[] = $filter;
         }
     }
 
-    /**
-     * {@inheritDoc}
-     */
-    public function pushProcessor(callable $callback): HandlerInterface
+    /** @throws void */
+    public function isHandling(LogRecord $record): bool
     {
-        array_unshift($this->processors, $callback);
-        return $this;
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    public function popProcessor(): callable
-    {
-        return array_shift($this->processors);
-    }
-
-    /**
-     * Checks whether the given record will be handled by this handler.
-     *
-     * This is mostly done for performance reasons, to avoid calling processors for nothing.
-     *
-     * Handlers should still check the record levels within handle(), returning false in isHandling()
-     * is no guarantee that handle() will not be called, and isHandling() might not be called
-     * for a given record.
-     *
-     * @param array<string, mixed> $record Partial log record containing only a level key
-     */
-    public function isHandling(array $record): bool
-    {
-        if ($record['level'] < $this->handlerLevel) {
+        if (!parent::isHandling($record)) {
             return false;
         }
 
-        if (isset($record['message'])) {
+        if (isset($record->message)) {
             // when record is fulfilled, try each filter
             foreach ($this->filters as $filter) {
-                if (!$filter($record, $this->handlerLevel)) {
+                if (!$filter($record, $this->level)) {
                     return false;
                 }
             }
@@ -141,46 +96,82 @@ class CallbackFilterHandler extends AbstractHandler implements ProcessableHandle
         return true;
     }
 
-    /**
-     * {@inheritdoc}
-     */
-    public function handle(array $record): bool
+    /** @throws RuntimeException */
+    public function handle(LogRecord $record): bool
     {
+        // The same logic as in FilterHandler
+
         if (!$this->isHandling($record)) {
             return false;
         }
 
-        // The same logic as in FingersCrossedHandler
-        if (!$this->handler instanceof HandlerInterface) {
-            $this->handler = ($this->handler)($record, $this);
-            if (!$this->handler instanceof HandlerInterface) {
-                throw new RuntimeException("The factory callable should return a HandlerInterface");
-            }
-        }
+        $record = $this->processRecord($record);
 
-        if ($this->processors) {
-            foreach ($this->processors as $processor) {
-                $record = ($processor)($record);
-            }
-        }
-
-        $this->handler->handle($record);
+        $this->getHandler($record)->handle($record);
 
         return false === $this->bubble;
     }
 
     /**
-     * {@inheritdoc}
+     * @param LogRecord[] $records
+     *
+     * @throws RuntimeException
      */
     public function handleBatch(array $records): void
     {
+        // The same logic as in FilterHandler
         $filtered = [];
+
         foreach ($records as $record) {
-            if ($this->isHandling($record)) {
-                $filtered[] = $record;
+            if (!$this->isHandling($record)) {
+                continue;
             }
+
+            $filtered[] = $record;
         }
 
-        $this->handler->handleBatch($filtered);
+        if (0 >= count($filtered)) {
+            return;
+        }
+
+        $this->getHandler($filtered[count($filtered) - 1])->handleBatch($filtered);
+    }
+
+    /**
+     * Return the nested handler
+     *
+     * If the handler was provided as a factory, this will trigger the handler's instantiation.
+     *
+     * @throws RuntimeException
+     */
+    public function getHandler(LogRecord | null $record = null): HandlerInterface
+    {
+        // The same logic as in FingersCrossedHandler
+
+        if (!$this->handler instanceof HandlerInterface) {
+            $handler = ($this->handler)($record, $this);
+
+            if (!$handler instanceof HandlerInterface) {
+                throw new RuntimeException('The factory Closure should return a HandlerInterface');
+            }
+
+            $this->handler = $handler;
+        }
+
+        return $this->handler;
+    }
+
+    /** @throws RuntimeException */
+    public function reset(): void
+    {
+        $this->resetProcessors();
+
+        $handler = $this->getHandler();
+
+        if (!($handler instanceof ResettableInterface)) {
+            return;
+        }
+
+        $handler->reset();
     }
 }
